+++ src/tab_delim.rs
@@ -58,7 +58,10 @@ pub(crate) fn load_tab_delimited_facts(
             var_drop_used,
             var_uses_region,
             var_drops_region,
-            var_initialized_on_exit,
+            child,
+            path_belongs_to_var,
+            initialized_at,
+            moved_out_at,
         }
     }
 }